Add SecuredController.
