+++ app/controllers/health_cards_controller.rb
@@ -3,7 +3,7 @@
 # require 'app/lib/covid_health_card_reporter'
 
 # HealthCardsController is the endpoint for download and issue of health cards
-class HealthCardsController < ApplicationController
+class HealthCardsController < SecuredController
   before_action :create_exporter, except: [:scan, :qr_contents, :upload]
   skip_before_action :verify_authenticity_token, only: [:create]
   after_action :set_cors_header, only: :create
