+++ app/controllers/patients_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # PatientsController manages patients via the Web UI
-class PatientsController < ApplicationController
+class PatientsController < SecuredController
   before_action :set_patient, only: %i[show edit update destroy]
   after_action :set_cors_header, only: :show
 
@@ -15,13 +15,7 @@ class PatientsController < ApplicationController
     respond_to do |format|
       format.html
       format.fhir_json { render json: @patient.to_json }
-      format.json do
-        if helpers.verify_token(request.headers)
-          render json: @patient.to_json
-        else
-          render json: { errors: ['Unauthorized code'] }, status: :unauthorized
-        end
-      end
+      format.json { render json: @patient.to_json }
     end
   end
