--- app/helpers/auth_helper.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-module AuthHelper
-  def get_access_token(headers)
-    headers['Authorization'].split.last if headers['Authorization'].present?
-  end
-
-  def verify_token(headers)
-    access_token = get_access_token headers
-    return true if access_token.nil?
-
-    # Verify token signature and retrieve payload
-    payload = JWT.decode(access_token, Rails.application.config.hc_key.key, true, { algorithm: 'ES256' }).first
-    # Check if token has expired
-    return false if convert_time_to_epoch(Time.current) > payload['exp']
-
-    true
-  rescue JWT::VerificationError, JWT::DecodeError
-    false
-  end
-end
